Fixed crucial bugs. New functions.
Throwing exceptions instead of std::cerr or std::cout
Fixed a few bugs that were with dividing by negative numbers and with certain operators. Fixed so many bugs.
This version is completely different from the previous version.
Added functions like prime number computing, hash functions, and efficient division and multiplication when the divisor/multiplier is a power of 2.